Tried and apparently failed to debug. I caught some bugs. They're still in there though. It's pretty rough. I genuinely am not sure we can even get inserts to work in RBT, if a simple BST got so complex so fast.
I tried to create a rough skeleton for a bullshit rbt visualizer, but even that will take some work. Maybe we could find a better RBT class to work with or create our own. That would not be so hard... with what we have, there is too much abstraction to easily draw the entire tree.